"Fixing" active flag status to match what the Java version is doing
The backend API doesn't allow an update to an existing spacecraft journey so the active status looks like it never completes in the python version. The java version simply sets this to "false" right out of the gate. Setting this value to "false" in the python version to match the java experience.